% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actxps-package.R
\docType{package}
\name{actxps-package}
\alias{actxps}
\alias{actxps-package}
\title{actxps: Create Actuarial Experience Studies: Prepare Data, Summarize Results, and Create Reports}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Experience studies are used by actuaries to explore historical experience across blocks of business and to inform assumption setting activities. This package provides functions for preparing data, creating studies, visualizing results, and beginning assumption development. Experience study methods, including exposure calculations, are described in: Atkinson & McGarry (2016) "Experience Study Calculations" \url{https://www.soa.org/49378a/globalassets/assets/files/research/experience-study-calculations.pdf}. The limited fluctuation credibility method used by the 'exp_stats()' function is described in: Herzog (1999, ISBN:1-56698-374-6) "Introduction to Credibility Theory".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattheaphy/actxps/}
  \item \url{https://mattheaphy.github.io/actxps/}
  \item Report bugs at \url{https://github.com/mattheaphy/actxps/issues}
}

}
\author{
\strong{Maintainer}: Matt Heaphy \email{mattrmattrs@gmail.com}

}
\keyword{internal}
