% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{add_transactions}
\alias{add_transactions}
\title{Add transactions to an experience study}
\usage{
add_transactions(
  .data,
  trx_data,
  col_pol_num = "pol_num",
  col_trx_date = "trx_date",
  col_trx_type = "trx_type",
  col_trx_amt = "trx_amt"
)
}
\arguments{
\item{.data}{a data frame with exposure-level records of type
\code{exposed_df}. Use \code{\link[=as_exposed_df]{as_exposed_df()}} to convert a data frame to an
\code{exposed_df} object if necessary.}

\item{trx_data}{a data frame containing transactions details}

\item{col_pol_num}{name of the column in \code{trx_data} containing the policy
number}

\item{col_trx_date}{name of the column in \code{trx_data} containing the
transaction date}

\item{col_trx_type}{name of the column in \code{trx_data} containing the
transaction type}

\item{col_trx_amt}{name of the column in \code{trx_data} containing the
transaction amount}
}
\value{
UPDATE ME
}
\description{
Attach summarized transactions to a data frame with
exposure-level records.
}
\details{
UPDATE ME
}
\examples{
expo <- expose_py(census_dat, "2019-12-31", target_status = "Surrender")
add_transactions(expo, withdrawals)

}
