% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_df_helpers.R
\name{as_exp_df}
\alias{as_exp_df}
\alias{is_exp_df}
\title{Termination summary helper functions}
\usage{
as_exp_df(
  x,
  expected = NULL,
  col_claims,
  col_exposure,
  target_status = NULL,
  start_date = NULL,
  end_date = NULL,
  credibility = FALSE,
  conf_level = 0.95,
  cred_r = 0.05,
  conf_int = FALSE
)

is_exp_df(x)
}
\arguments{
\item{x}{An object. For \code{as_exp_df()}, \code{x} must be a data frame.}

\item{expected}{A character vector containing column names in x with
expected values}

\item{col_claims}{Optional. Name of the column in \code{x} containing claims. The
assumed default is "claims".}

\item{col_exposure}{Optional. Name of the column in \code{x} containing exposures.
The assumed default is "exposure".}

\item{target_status}{Character vector of target status values. Default value
= \code{NULL}.}

\item{start_date}{Experience study start date. Default value = 1900-01-01.}

\item{end_date}{Experience study end date}

\item{credibility}{If \code{TRUE}, future calls to \code{\link[=summary]{summary()}} will include
partial credibility weights and credibility-weighted termination rates.}

\item{conf_level}{Confidence level used for the Limited Fluctuation
credibility method and confidence intervals}

\item{cred_r}{Error tolerance under the Limited Fluctuation credibility
method}

\item{conf_int}{If \code{TRUE}, future calls to \code{\link[=summary]{summary()}} will include
confidence intervals around the observed termination rates and any
actual-to-expected ratios.}
}
\value{
For \code{is_exp_df()}, a length-1 logical vector. For \code{as_exp_df()},
an \code{exp_df} object.
}
\description{
Test for and coerce to the \code{exp_df} class.
}
\details{
\code{is_exp_df()} will return \code{TRUE} if \code{x} is an \code{exp_df} object.

\code{as_exp_df()} will coerce a data frame to an \code{exp_df} object if that
data frame has columns for exposures and claims.

\code{as_exp_df()} is most useful for converting existing aggregate summaries of
experience where individual policy information is not available.
}
\seealso{
exp_stats
}
