% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{autoplot_exp}
\alias{autoplot_exp}
\alias{autoplot.exp_df}
\alias{autoplot.trx_df}
\title{Plot experience study results}
\usage{
\method{autoplot}{exp_df}(
  object,
  ...,
  x = NULL,
  y = NULL,
  color = NULL,
  mapping,
  second_axis = FALSE,
  second_y = NULL,
  scales = "fixed",
  geoms = c("lines", "bars", "points"),
  y_labels = scales::label_percent(accuracy = 0.1),
  second_y_labels = scales::label_comma(accuracy = 1),
  y_log10 = FALSE,
  conf_int_bars = FALSE
)

\method{autoplot}{trx_df}(
  object,
  ...,
  x = NULL,
  y = NULL,
  color = NULL,
  mapping,
  second_axis = FALSE,
  second_y = NULL,
  scales = "fixed",
  geoms = c("lines", "bars", "points"),
  y_labels = scales::label_percent(accuracy = 0.1),
  second_y_labels = scales::label_comma(accuracy = 1),
  y_log10 = FALSE,
  conf_int_bars = FALSE
)
}
\arguments{
\item{object}{An object of class \code{exp_df} created by the
function \code{\link[=exp_stats]{exp_stats()}} or an object of class \code{trx_df} created by the function
\code{\link[=trx_stats]{trx_stats()}}.}

\item{...}{Faceting variables passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}

\item{x}{An unquoted column name in \code{object} or expression to use as the \code{x}
variable.}

\item{y}{An unquoted column name in \code{object} or expression to use as the
\code{y} variable. If unspecified, \code{y} will default to the observed termination
rate (\code{q_obs}) for \code{exp_df} objects and the observed utilization rate
(\code{trx_util}) for \code{trx_df} objects.}

\item{color}{An unquoted column name in \code{object} or expression to use as the
\code{color} and \code{fill} variables.}

\item{mapping}{Aesthetic mapping passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}. NOTE: If
\code{mapping} is supplied, the \code{x}, \code{y}, and \code{color} arguments will be ignored.}

\item{second_axis}{Logical. If \code{TRUE}, the variable specified by
\code{second_y} (default = exposure) is plotted on a second
y-axis using an area geometry.}

\item{second_y}{An unquoted column name in \code{object} to use as the \code{y}
variable on the second y-axis. If unspecified, this will default to
\code{exposure}.}

\item{scales}{The \code{scales} argument passed to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}

\item{geoms}{Type of geometry. If "lines" is passed, the plot will
display lines and points. If "bars", the plot will display bars. If "points",
the plot will display points only.}

\item{y_labels}{Label function passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}.}

\item{second_y_labels}{Same as \code{y_labels}, but for the second y-axis.}

\item{y_log10}{If \code{TRUE}, the y-axes are plotted on a log-10 scale.}

\item{conf_int_bars}{If \code{TRUE}, confidence interval error bars are included
in the plot. This option is only available for termination rates and
actual-to-expected ratios.}
}
\value{
a \code{ggplot} object
}
\description{
Plot experience study results
}
\details{
If no aesthetic map is supplied, the plot will use the first
grouping variable in \code{object} on the x axis and \code{q_obs} on the y
axis. In addition, the second grouping variable in \code{object} will be
used for color and fill.

If no faceting variables are supplied, the plot will use grouping
variables 3 and up as facets. These variables are passed into
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}. Specific to \code{trx_df} objects, transaction
type (\code{trx_type}) will also be added as a faceting variable.
}
\examples{

study_py <- expose_py(census_dat, "2019-12-31", target_status = "Surrender")

study_py <- study_py |>
  add_transactions(withdrawals)

exp_res <- study_py |> group_by(pol_yr) |> exp_stats()
autoplot(exp_res)

trx_res <- study_py |> group_by(pol_yr) |> trx_stats()
autoplot(trx_res)

}
\seealso{
\code{\link[=plot_termination_rates]{plot_termination_rates()}}, \code{\link[=plot_actual_to_expected]{plot_actual_to_expected()}}
}
