% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{.actxps_plot_theme}
\alias{.actxps_plot_theme}
\title{Plotting theme for vignettes}
\usage{
.actxps_plot_theme()
}
\value{
No return value. This function is called for its side effects.
}
\description{
This is an internal function used to set a plotting theme in vignettes
and articles.
}
