% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_explore.R
\name{exp_explore}
\alias{exp_explore}
\title{Interactively explore experience data}
\usage{
exp_explore(dat, predictors = names(dat))
}
\arguments{
\item{dat}{An \code{exposed_df} object}
}
\value{
\code{NULL}
}
\description{
Launch a shiny application to interactively explore drivers of
experience.

\strong{temp} details here. note filters and data visualization, also removal
of reserved column names from predictors. also explain exposed_df importance
}
\examples{
toy_census |> expose("2020-12-31", target_status = "Surrender")

\dontrun{1}

}
