% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_explore.R
\name{exp_shiny}
\alias{exp_shiny}
\title{Interactively explore experience data}
\usage{
exp_shiny(
  dat,
  predictors = names(dat),
  expected = stringr::str_subset(names(dat), "expected")
)
}
\arguments{
\item{dat}{An \code{exposed_df} object}

\item{predictors}{A character vector of independent variables in \code{dat} to
include in the shiny app.}

\item{expected}{A character vector of expected values in \code{dat} to include
in the shiny app.}
}
\value{
\code{NULL}
}
\description{
Launch a shiny application to interactively explore drivers of
experience.

\strong{temp} details here. note filters and data visualization, also removal
of reserved column names from predictors. also explain exposed_df importance
}
\examples{

\dontrun{1}

}
