% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_stats.R
\name{exp_stats}
\alias{exp_stats}
\title{Summarize experience study records}
\usage{
exp_stats(.data, expected, target_status)
}
\arguments{
\item{.data}{a data frame with exposure-level records}

\item{expected}{a character vector of expected values}

\item{target_status}{a character vector of target status values}
}
\value{
something useful
}
\description{
update me
}
\details{
add more details here

If \code{target_status} isn't provided, all status values except the first level will be used. This method
implicitly assumes that the first level is a default active status.
}
\examples{
1

}
