% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose.R
\name{expose}
\alias{expose}
\title{Create exposoure records from census records}
\usage{
expose(dat)
}
\arguments{
\item{dat}{a data frame with census-level records}
}
\value{
something useful
}
\description{
update me
}
\details{
add details here
}
\examples{
1
}
