% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_special.R
\name{plot_special}
\alias{plot_special}
\alias{plot_termination_rates}
\title{Plot observed and expected termination rates}
\usage{
plot_termination_rates(object, ...)
}
\arguments{
\item{object}{An object of class \code{exp_df} created by the function
\code{\link[=exp_stats]{exp_stats()}}.}

\item{...}{Additional arguments passed to \code{autoplot.exp_df}.}
}
\value{
a \code{ggplot} object
}
\description{
Plot observed and expected termination rates
}
\details{
Create a plot of all observed and expected termination rates.
}
\examples{

study_py <- expose_py(census_dat, "2019-12-31", target_status = "Surrender")
expected_table <- c(seq(0.005, 0.03, length.out = 10), 0.2, 0.15, rep(0.05, 3))

study_py <- study_py |>
  mutate(expected_1 = expected_table[pol_yr],
         expected_2 = ifelse(inc_guar, 0.015, 0.03))

exp_res <- study_py |> group_by(pol_yr) |>
  exp_stats(expected = c("expected_1", "expected_2"))
plot_termination_rates(exp_res)

}
