% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_basic_2012_g2.R
\docType{data}
\name{qx_iamb}
\alias{qx_iamb}
\alias{scale_g2}
\title{2012 Individual Annuity Mortality Table and Projection Scale G2}
\format{
For the 2012 IAMB table, a data frame with 242 rows and 3 columns:
\describe{
\item{age}{attained age}
\item{qx}{mortality rate}
\item{gender}{Female or Male}
}

For the Project Scale G2 table, a data frame with 242 rows and 3 columns:
\describe{
\item{age}{attained age}
\item{mi}{mortality improvment rate}
\item{gender}{Female or Male}
}
}
\source{
\itemize{
\item \url{https://mort.soa.org/}
\item \url{https://www.actuary.org/sites/default/files/files/publications/Payout_Annuity_Report_09-28-11.pdf}
}
}
\usage{
qx_iamb

scale_g2
}
\description{
Mortality rates and mortality improvement rates from the 2012 Individual
Annuity Mortality Basic (IAMB) Table and Project Scale G2.
}
\keyword{datasets}
