% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_expose.R
\name{step_expose}
\alias{step_expose}
\title{Create exposure records in a \code{recipes} step}
\usage{
step_expose(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  end_date,
  start_date = as.Date("1900-01-01"),
  target_status = NULL,
  cal_expo = FALSE,
  expo_length = "year",
  col_names = list(col_pol_num = "pol_num", col_status = "status", col_issue_date =
    "issue_date", col_term_date = "term_date"),
  default_status,
  skip = TRUE,
  id = recipes::rand_id("expose")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step. See \code{\link[recipes:selections]{selections()}} for more details.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{end_date}{experience study end date}

\item{start_date}{experience study start date. Default value = 1900-01-01.}

\item{target_status}{character vector of target status values. Default value = \code{NULL}.}

\item{cal_expo}{set to TRUE for calendar year exposures. Otherwise policy year exposures are used.}

\item{expo_length}{exposure period length}

\item{col_names}{A named list of column names passed to the \code{col_pol_num},
\code{col_status}, \code{col_issue_date}, and \code{col_term_date} arguments of \code{expose()}.}

\item{default_status}{optional scalar character representing the default active status code}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake]{bake()}}? While all operations are baked
when \code{\link[recipes:prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\description{
\code{step_expose} creates a \emph{specification} of a recipe step that will convert
a data frame of census-level records to exposure-level records.
}
\examples{

1

}
\seealso{
\code{\link[=expose]{expose()}}
}
