% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_census.R
\docType{data}
\name{toy_census}
\alias{toy_census}
\title{Toy policy census data}
\format{
A data frame with 3 rows and 4 columns:
\describe{
\item{pol_num}{policy number}
\item{status}{policy status}
\item{issue_date}{issue date}
\item{term_date}{termination date}
}
}
\usage{
toy_census
}
\description{
A tiny dataset containing 3 policies: one active, one terminated due to
death, and one terminated due to surrender.
}
\keyword{datasets}
