% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{autoplot.exp_df}
\alias{autoplot.exp_df}
\title{Plot experience study results}
\usage{
\method{autoplot}{exp_df}(object, ...)
}
\arguments{
\item{object}{An object of class \code{exp_df} usually created by the
function \code{exp_stats}.}

\item{...}{Not currently used.}
}
\description{
Plot experience study results
}
