% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{autotable}
\alias{autotable}
\alias{autotable.exp_df}
\title{Tabular experience study summary}
\usage{
autotable(object, ...)

\method{autotable}{exp_df}(object, fontsize = 100, decimals = 1, ...)
}
\arguments{
\item{object}{An object of class \code{exp_df} usually created by the
function \code{exp_stats()}.}

\item{...}{Additional arguments passed to \code{gt::gt()}.}

\item{fontsize}{Font size percentage multiplier.}

\item{decimals}{Number of decimals to display for percentages}
}
\value{
a \code{gt} object
}
\description{
\code{autotable} is a generic function used to create a table
from an object of a particular class. Tables are constructed using the
\code{gt} package.

\code{autotable.exp_df} is used to convert experience study results to a
presentation-friendly format.
}
