% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose.R
\name{expose}
\alias{expose}
\alias{expose_cy}
\title{Create exposure records from census records}
\usage{
expose(
  .data,
  end_date,
  start_date = as.Date("1900-01-01"),
  target_status = NULL,
  col_pol_num = "pol_num",
  col_status = "status",
  col_issue_date = "issue_date",
  col_term_date = "term_date",
  default_status
)

expose_cy(
  .data,
  end_date,
  start_date = as.Date("1900-01-01"),
  target_status = NULL,
  col_pol_num = "pol_num",
  col_status = "status",
  col_issue_date = "issue_date",
  col_term_date = "term_date",
  default_status
)
}
\arguments{
\item{.data}{a data frame with census-level records}

\item{end_date}{experience study end date}

\item{start_date}{experience study start date. Default value = 1900-01-01.}

\item{target_status}{character vector of target status values. Default value = \code{NULL.}}

\item{col_pol_num}{name of the column in \code{.data} containing the policy number}

\item{col_status}{name of the column in \code{.data} containing the policy status}

\item{col_issue_date}{name of the column in \code{.data} containing the issue date}

\item{col_term_date}{name of the column in \code{.data} containing the termination date}

\item{default_status}{optional scalar character representing the default active status code.}
}
\value{
A tibble with class \code{exposed_df}, \code{tbl_df}, \code{tbl},
and \code{data.frame}. The results include all existing columns in
\code{.data} plus new columns for exposures and observation periods.

For policy year and policy month exposures, any calendar-based observation
periods represent the beginning of the policy year or policy month. For
example, using a policy year exposure basis, assume that for a particular
record the policy year (\code{pol_yr}) is 3 and the calendar year
(\code{cal_yr}) is 2022. This means that it was 2022 at the start of policy
year 3.
}
\description{
Convert a data frame of census-level records to exposure-level
records.
}
\details{
Census-level data refers to a data set wherein there is one row
per unique policy. Exposure-level data expands census-level data such that
there is one record per policy per observation period. Observation periods
could be any meaningful period of time such as a policy year, policy month,
calendar year, calendar quarter, calendar month, etc.

\code{target_status} is used in the calculation of exposures. The annual
exposure method is applied, which allocates a full period of exposure for
any statuses in \code{target_status}. For all other statuses, new entrants
and exits are partially exposed based on the time elapsed in the observation
period. This method is consistent with the Balducci Hypothesis, which assumes
that the probability of termination is proportionate to the time elapsed
in the observation period. If the annual exposure method isn't desired,
\code{target_status} can be ignored. In this case, partial exposures are
always applied regardless of status.

\code{default_status} is used to indicate the default active status that
should be used when exposure records are created. If left blank, then the
first status level will be assumed to be the default active status.
}
\examples{
toy_census |> expose("2020-12-31", target_status = "Surrender")

\dontrun{
census_dat |>
    expose("2019-12-31", target_status = "Surrender")}

}
\references{
https://www.soa.org/49378a/globalassets/assets/files/research/experience-study-calculations.pdf
}
