% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose_split.R
\name{expose_split}
\alias{expose_split}
\title{Split calendar year exposures by policy year}
\usage{
expose_split(dat)
}
\arguments{
\item{dat}{An \code{exposed_df} object with calendar year exposures.}
}
\value{
A tibble with class \code{split_exposed_df}, \code{exposed_df}, \code{tbl_df},
\code{tbl}, and \code{data.frame}. The results include all columns in \code{dat} except that
\code{exposure} has been renamed to \code{exposure_cal}. Additional columns include:
\itemize{
\item \code{exposure_pol} - policy year exposures
\item \code{pol_yr} - policy year
\item \code{piece} - a factor containing 2 levels: "pre_anniv" (pre-anniversary
records) and "post_anniv" (post-anniversary records)
}
}
\description{
Convert a data frame of calendar year exposure-level records
into a split exposed data frame that divides each calendar year into two
pieces based on policy years: a pre-anniversary record and a post-anniversary
record.
}
\details{
\code{dat} must be an \code{exposed_df} with calendar year exposure records.
Calendar year exposures are created by the function \code{\link[=expose_cy]{expose_cy()}} (or
\code{\link[=expose]{expose()}} when \code{expo_length = "year"} and \code{cal_expo = TRUE}).
}
\examples{
toy_census |> expose_cy("2022-12-31") |> expose_split()

}
\seealso{
\code{\link[=expose]{expose()}}
}
