% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{groups}
\alias{group_by}
\alias{ungroup}
\alias{filter}
\alias{arrange}
\alias{mutate}
\alias{select}
\alias{slice}
\alias{rename}
\alias{relocate}
\alias{left_join}
\alias{right_join}
\alias{inner_join}
\alias{full_join}
\alias{semi_join}
\alias{anti_join}
\alias{prep}
\alias{bake}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:filter-joins]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr:mutate-joins]{full_join}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr:group_data]{groups}}, \code{\link[dplyr:mutate-joins]{inner_join}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:mutate-joins]{right_join}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:filter-joins]{semi_join}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:group_by]{ungroup}}}

  \item{generics}{\code{\link[generics]{tidy}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{recipes}{\code{\link[recipes]{bake}}, \code{\link[recipes]{prep}}}
}}

