% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trx_stats.R
\name{trx_stats}
\alias{trx_stats}
\alias{summary.trx_df}
\title{Summarize transactions and utilization rates}
\usage{
trx_stats(.data, trx_types, full_exposures = TRUE)

\method{summary}{trx_df}(object, ...)
}
\arguments{
\item{.data}{a data frame with exposure-level records of type
\code{exposed_df} with transaction data attached. If necessary, use
\code{\link[=as_exposed_df]{as_exposed_df()}} to convert a data frame to an \code{exposed_df} object, and use
\code{\link[=add_transactions]{add_transactions()}} to attach transactions to an \code{exposed_df} object.}

\item{trx_types}{A character vector of transaction types to include in the
output. If none is provided, all available transaction types in \code{.data}
will be used.}

\item{full_exposures}{If \code{TRUE} (default), partially exposed records will
be excluded from \code{data}.}

\item{object}{an \code{trx_df} object}

\item{...}{groups to retain after \code{summary()} is called}
}
\value{
UPDATE ME
}
\description{
Create a summary data frame of transaction counts, amounts,
and utilization rates.
}
\details{
UPDATE ME
}
\examples{
#UPDATE ME

}
